se valida la elimienación de projects y concepts

+++ src/expenses/services/concepts.services.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, ConflictException, HttpException, Injectable, InternalServerErrorException, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Concept } from '../entities/concept.entity';
@@ -28,11 +28,11 @@ export class ConceptsService {
         // return await this.conceptsRepository.find();
         try {
             const concepts = await this.conceptsRepository
-            .createQueryBuilder('c')
-            .leftJoin('projects', 'p', 'p.id = c.project_id')
-            .select(['c.*', 'p.name as project_name'])
-            .where('c.company_id = :companyId', { companyId })
-            .getRawMany();
+                .createQueryBuilder('c')
+                .leftJoin('projects', 'p', 'p.id = c.project_id')
+                .select(['c.*', 'p.name as project_name'])
+                .where('c.company_id = :companyId', { companyId })
+                .getRawMany();
 
             return concepts;
         } catch (error) {
@@ -47,7 +47,7 @@ export class ConceptsService {
     }
 
     async findByProjectId(id: number, companyId: number, select?: string) {
-        
+
         try {
             const projects = await this.conceptsRepository
                 .createQueryBuilder('concept')
@@ -76,7 +76,15 @@ export class ConceptsService {
         }
     }
 
-    remove(id: number) {
-        return this.conceptsRepository.delete(id);
+    async remove(id: number) {
+        try {
+            const concept = await this.conceptsRepository.query(`select id from expenses where concept = $1 limit 1`, [id]);
+            if (concept && concept.length > 0) {
+                throw new ConflictException('Concept not deleted because it is associated with existing expenses');
+            }
+            return this.conceptsRepository.delete(id);
+        } catch (error) {
+            throw new HttpException({ message: error.message }, error.status ? error.status : 500);
+        }
     }
 }
